requirements.txt
asgiref==3.8.1
Django==3.1.12
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djongo==1.3.7
PyJWT==2.9.0
pymongo==3.11.4
pytz==2025.1
sqlparse==0.2.4
tzdata==2025.1
